test(cloud): satisfy changed-file coverage
