+++ scripts/security/coverage-subprocess-sources.txt
@@ -26,3 +26,6 @@ packages/eliza-hub/services/merge-steward/src/cli.js
 
 # ACP stdio entrypoint is exercised via spawned runtime contract tests; importing it in-process starts the protocol server.
 packages/examples/code/src/acp.ts
+
+# The cloud-shared package test wrapper is exercised by a spawned real-process contract suite; importing it starts the test runner.
+packages/cloud/shared/scripts/run-bun-tests.mjs
